Add hello_world.kt
This is a simple Kotlin program that prints "Hello, World!" to the console. Kotlin is a modern programming language that can be run easily in online Kotlin compilers or environments like IntelliJ IDEA.

How to run :

Open an online Kotlin compiler like Kotlin Playground or use a local development environment than Paste the code and click Run.